Update incubation guidelines for server relevance (#36)
This PR updates the incubation process to be more specific about the SSWG's goals for the package index. The main idea I'm attempting to present here is that packages submitted to the SSWG should demonstrate a meaningful impact for server use cases specifically. The SSWG's package index is not meant to be a general purpose Swift package index.